pyproject.toml
[project]
name = "llm-sla"
version = "0.1.0"
description = "This is a LLM SLA test appilcation"
authors = "yicheng"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "numpy>=2.2.4",
    "pytest-playwright>=0.7.0",
    "tiktoken>=0.9.0",
]
